requirements.txt
livekit-agents[deepgram,openai,silero]>=0.12.0
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-silero
openai>=1.0.0
qdrant-client>=1.9.0
langgraph>=0.2.0
langchain-core>=0.2.0
python-dotenv
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
tzdata
